refactor: break secret_env_plan <-> env_materialization_plan cycle
env_materialization_plan and secret_env_plan formed a mutual dependency:
secret_env_plan holds an EnvMaterializationPlan field (structural), while
EnvMaterializationPlan::from_secret_env_plan read a SecretEnvPlan (a conversion
constructor). That back-edge blocked both modules from moving into the
homeboy-lab-contract foundation crate.

Invert the conversion: replace EnvMaterializationPlan::from_secret_env_plan with
SecretEnvPlan::to_env_materialization_plan (identical body), so env_materialization_plan
no longer references secret_env_plan at all. secret_env_plan reuses its own
existing normalize_names helper. Move the conversion's test alongside the method.

Result: env_materialization_plan -> secret_env_plan references drop to 0
(production and test); the dependency flows one way. Cut 4b toward the
homeboy-lab-contract foundation crate.
